Deploy/software/requirements.txt
boto3==1.24.28
numpy==1.21.5
Pillow==9.5.0
Requests==2.30.0
torch==1.12.1
torchvision==0.13.1
